feat: move logs printing to module
